Multiple clean ups and minor additions 🧹
- Make `AnalyticsTracker` extend `AnyObject`

- Update `: class` to `: AnyObject` on protocols.

- Rename and reorganize some files.

- Remove leftover copyright headers on files.

- Add protocol extension on `EmptyExternalResource`, when
`External == Data`.

- Handle 205 HTTP status codes on `URLSessionNetworkStack`, which also
return an empty body on success.
