+++ Sources/PerformanceMetrics/Trackers/PerformanceMetricsTracker.swift
@@ -2,7 +2,7 @@ import Foundation
 import Result
 
 /// A type that tracks performance metrics.
-public protocol PerformanceMetricsTracker: class {
+public protocol PerformanceMetricsTracker: AnyObject {
 
     /// A metric unique tag.
     typealias Tag = PerformanceMetrics.Tag
