+++ Sources/Persistence/CoreData/CoreDataEntity.swift
@@ -1,6 +1,6 @@
 import CoreData
 
-public protocol CoreDataEntity: class, NSFetchRequestResult {
+public protocol CoreDataEntity: AnyObject & NSFetchRequestResult {
     static var entityName: String { get }
 }
