+++ Tests/AlicerceTests/Extensions/URLRequestTestCase.swift
@@ -1,5 +1,3 @@
-// Copyright © 2018 Mindera. All rights reserved.
-
 import XCTest
 @testable import Alicerce
